split out rendering helpers and sprite constants
